The modal overlay covers only the current viewing area. If the page is really long, that is not being covered. The user could scroll down to an “unprotected” area and click the background link without closing the modal 1st.
Also, the modal opens up at the top. The user could scroll down to the point where the modal initiating button/link was partially visible and click on it. That would show the modal at top (could also be so up that it may be out of view of the user).